LayoutTests/imported/w3c/web-platform-tests/css/css-values/attr-all-types-expected.txt


FAIL CSS Values and Units Test: attr assert_equals: Value 'attr(data-foo string)', where 'data-foo=abc' should be valid for the property 'content'. expected "\"abc\"" but got "normal"
FAIL CSS Values and Units Test: attr 1 assert_equals: Value 'attr(data-foo string)', where 'data-foo=attr(data-foo)' should be valid for the property 'content'. expected "\"attr(data-foo)\"" but got "normal"
FAIL CSS Values and Units Test: attr 2 assert_equals: Value 'attr(data-foo ident)', where 'data-foo=anim' should be valid for the property 'animation-name'. expected "anim" but got "none"
FAIL CSS Values and Units Test: attr 3 assert_equals: Value 'attr(data-foo ident, anim-fallback)', where 'data-foo="anim"' should be valid for the property 'animation-name'. expected "anim-fallback" but got "none"
FAIL CSS Values and Units Test: attr 4 assert_equals: Value 'attr(data-foo ident, anim-fallback)', where 'data-foo=initial' should be valid for the property 'animation-name'. expected "anim-fallback" but got "none"
FAIL CSS Values and Units Test: attr 5 assert_equals: Value 'attr(data-foo color)', where 'data-foo=red' should be valid for the property 'background-color'. expected "rgb(255, 0, 0)" but got "rgba(0, 0, 0, 0)"
FAIL CSS Values and Units Test: attr 6 assert_equals: Value 'attr(data-foo color)', where 'data-foo=#ff0099aa' should be valid for the property 'background-color'. expected "rgba(255, 0, 153, 0.667)" but got "rgba(0, 0, 0, 0)"
FAIL CSS Values and Units Test: attr 7 assert_equals: Value 'attr(data-foo color, red)', where 'data-foo=10' should be valid for the property 'background-color'. expected "rgb(255, 0, 0)" but got "rgba(0, 0, 0, 0)"
FAIL CSS Values and Units Test: attr 8 assert_equals: Value 'attr(data-foo color, green)', where 'data-foo=1000px' should be valid for the property 'background-color'. expected "rgb(0, 128, 0)" but got "rgba(0, 0, 0, 0)"
FAIL CSS Values and Units Test: attr 9 assert_equals: Value 'attr(data-foo color, green)', where 'data-foo=rgb(255, 0, 0)' should be valid for the property 'background-color'. expected "rgb(0, 128, 0)" but got "rgba(0, 0, 0, 0)"
FAIL CSS Values and Units Test: attr 10 assert_equals: Value 'attr(data-foo number)', where 'data-foo=10' should be valid for the property 'font-weight'. expected "10" but got "400"
FAIL CSS Values and Units Test: attr 11 assert_equals: Value 'attr(data-foo number, 30)', where 'data-foo=10px' should be valid for the property 'font-weight'. expected "30" but got "400"
FAIL CSS Values and Units Test: attr 12 assert_equals: Value 'attr(data-foo number, calc(10 + 20))', where 'data-foo=10px' should be valid for the property 'font-weight'. expected "30" but got "400"
FAIL CSS Values and Units Test: attr 13 assert_equals: Value 'attr(data-foo percentage)', where 'data-foo=10%' should be valid for the property 'font-size'. expected "9px" but got "16px"
FAIL CSS Values and Units Test: attr 14 assert_equals: Value 'attr(data-foo percentage, 10px)', where 'data-foo=abc' should be valid for the property 'font-size'. expected "10px" but got "16px"
FAIL CSS Values and Units Test: attr 15 assert_equals: Value 'attr(data-foo percentage, abc)', where 'data-foo=10' should be valid for the property '--x'. expected "abc" but got "attr(data-foo percentage, abc)"
FAIL CSS Values and Units Test: attr 16 assert_equals: Value 'attr(data-foo length)', where 'data-foo=10px' should be valid for the property 'width'. expected "10px" but got "784px"
FAIL CSS Values and Units Test: attr 17 assert_equals: Value 'attr(data-foo length, red)', where 'data-foo=10px' should be valid for the property 'width'. expected "10px" but got "784px"
FAIL CSS Values and Units Test: attr 18 assert_equals: Value 'attr(data-foo length, 42px)', where 'data-foo=calc(1px + 3px)' should be valid for the property 'width'. expected "42px" but got "784px"
PASS CSS Values and Units Test: attr 19
PASS CSS Values and Units Test: attr 20
FAIL CSS Values and Units Test: attr 21 assert_equals: Value 'attr(data-foo angle, italic)', where 'data-foo=30' should be valid for the property 'font-style'. expected "italic" but got "normal"
FAIL CSS Values and Units Test: attr 22 assert_equals: Value 'attr(data-foo time)', where 'data-foo=10ms' should be valid for the property 'transition-duration'. expected "0.01s" but got "0s"
FAIL CSS Values and Units Test: attr 23 assert_equals: Value 'attr(data-foo time, 30s)', where 'data-foo=10m' should be valid for the property 'transition-duration'. expected "30s" but got "0s"
FAIL CSS Values and Units Test: attr 24 assert_equals: Value 'attr(data-foo time, calc(10s + 20s))', where 'data-foo=10m' should be valid for the property 'transition-duration'. expected "30s" but got "0s"
FAIL CSS Values and Units Test: attr 25 assert_equals: Value '30px attr(data-foo flex)', where 'data-foo=1fr' should be valid for the property 'grid-template-columns'. expected "30px 1fr" but got "none"
FAIL CSS Values and Units Test: attr 26 assert_equals: Value 'attr(data-foo flex, 3fr)', where 'data-foo=1fr 1fr' should be valid for the property 'grid-template-columns'. expected "3fr" but got "none"
FAIL CSS Values and Units Test: attr 27 assert_equals: Value 'attr(data-foo flex, calc(1px + 2px))', where 'data-foo=10px' should be valid for the property 'grid-template-columns'. expected "3px" but got "none"
FAIL CSS Values and Units Test: attr 28 assert_equals: Value 'attr(data-foo px)', where 'data-foo=10' should be valid for the property 'height'. expected "10px" but got "0px"
FAIL CSS Values and Units Test: attr 29 assert_equals: Value 'calc(attr(data-foo px) + 1px)', where 'data-foo=10' should be valid for the property 'width'. expected "11px" but got "784px"
FAIL CSS Values and Units Test: attr 30 assert_equals: Value 'attr(data-foo px) 11px', where 'data-foo=10' should be valid for the property '--x'. expected "10px 11px" but got "attr(data-foo px) 11px"
FAIL CSS Values and Units Test: attr 31 assert_equals: Value 'attr(data-foo fr)', where 'data-foo=10' should be valid for the property 'grid-template-columns'. expected "10fr" but got "none"
FAIL CSS Values and Units Test: attr 32 assert_equals: Value 'attr(data-foo fr, 3fr)', where 'data-foo=10fr' should be valid for the property 'grid-template-columns'. expected "3fr" but got "none"
FAIL CSS Values and Units Test: attr 33 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3em' should be valid for the property 'width'. expected "48px" but got "784px"
FAIL CSS Values and Units Test: attr 34 assert_equals: Value 'attr(data-foo em)', where 'data-foo=3' should be valid for the property 'width'. expected "48px" but got "784px"
FAIL CSS Values and Units Test: attr 35 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3ex' should be valid for the property 'width'. expected "21.53125px" but got "784px"
FAIL CSS Values and Units Test: attr 36 assert_equals: Value 'attr(data-foo ex)', where 'data-foo=3' should be valid for the property 'width'. expected "21.53125px" but got "784px"
FAIL CSS Values and Units Test: attr 37 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3cap' should be valid for the property 'width'. expected "31.75px" but got "784px"
FAIL CSS Values and Units Test: attr 38 assert_equals: Value 'attr(data-foo cap)', where 'data-foo=3' should be valid for the property 'width'. expected "31.75px" but got "784px"
FAIL CSS Values and Units Test: attr 39 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3ch' should be valid for the property 'width'. expected "24px" but got "784px"
FAIL CSS Values and Units Test: attr 40 assert_equals: Value 'attr(data-foo ch)', where 'data-foo=3' should be valid for the property 'width'. expected "24px" but got "784px"
FAIL CSS Values and Units Test: attr 41 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3ic' should be valid for the property 'width'. expected "48px" but got "784px"
FAIL CSS Values and Units Test: attr 42 assert_equals: Value 'attr(data-foo ic)', where 'data-foo=3' should be valid for the property 'width'. expected "48px" but got "784px"
FAIL CSS Values and Units Test: attr 43 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3rem' should be valid for the property 'width'. expected "48px" but got "784px"
FAIL CSS Values and Units Test: attr 44 assert_equals: Value 'attr(data-foo rem)', where 'data-foo=3' should be valid for the property 'width'. expected "48px" but got "784px"
FAIL CSS Values and Units Test: attr 45 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3lh' should be valid for the property 'width'. expected "54px" but got "784px"
FAIL CSS Values and Units Test: attr 46 assert_equals: Value 'attr(data-foo lh)', where 'data-foo=3' should be valid for the property 'width'. expected "54px" but got "784px"
FAIL CSS Values and Units Test: attr 47 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3rlh' should be valid for the property 'width'. expected "54px" but got "784px"
FAIL CSS Values and Units Test: attr 48 assert_equals: Value 'attr(data-foo rlh)', where 'data-foo=3' should be valid for the property 'width'. expected "54px" but got "784px"
FAIL CSS Values and Units Test: attr 49 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3vw' should be valid for the property 'width'. expected "24px" but got "784px"
FAIL CSS Values and Units Test: attr 50 assert_equals: Value 'attr(data-foo vw)', where 'data-foo=3' should be valid for the property 'width'. expected "24px" but got "784px"
FAIL CSS Values and Units Test: attr 51 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3vh ' should be valid for the property 'width'. expected "18px" but got "784px"
FAIL CSS Values and Units Test: attr 52 assert_equals: Value 'attr(data-foo vh )', where 'data-foo=3' should be valid for the property 'width'. expected "18px" but got "784px"
FAIL CSS Values and Units Test: attr 53 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3vi' should be valid for the property 'width'. expected "24px" but got "784px"
FAIL CSS Values and Units Test: attr 54 assert_equals: Value 'attr(data-foo vi)', where 'data-foo=3' should be valid for the property 'width'. expected "24px" but got "784px"
FAIL CSS Values and Units Test: attr 55 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3vb' should be valid for the property 'width'. expected "18px" but got "784px"
FAIL CSS Values and Units Test: attr 56 assert_equals: Value 'attr(data-foo vb)', where 'data-foo=3' should be valid for the property 'width'. expected "18px" but got "784px"
FAIL CSS Values and Units Test: attr 57 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3vmin' should be valid for the property 'width'. expected "18px" but got "784px"
FAIL CSS Values and Units Test: attr 58 assert_equals: Value 'attr(data-foo vmin)', where 'data-foo=3' should be valid for the property 'width'. expected "18px" but got "784px"
FAIL CSS Values and Units Test: attr 59 assert_equals: Value 'attr(data-foo length)', where 'data-foo=3vmax' should be valid for the property 'width'. expected "24px" but got "784px"
FAIL CSS Values and Units Test: attr 60 assert_equals: Value 'attr(data-foo vmax)', where 'data-foo=3' should be valid for the property 'width'. expected "24px" but got "784px"
PASS CSS Values and Units Test: attr 61
PASS CSS Values and Units Test: attr 62
PASS CSS Values and Units Test: attr 63
PASS CSS Values and Units Test: attr 64
PASS CSS Values and Units Test: attr 65
PASS CSS Values and Units Test: attr 66
PASS CSS Values and Units Test: attr 67
PASS CSS Values and Units Test: attr 68
FAIL CSS Values and Units Test: attr 69 assert_equals: Value 'attr(data-foo time)', where 'data-foo=3ms' should be valid for the property 'transition-duration'. expected "0.003s" but got "0s"
FAIL CSS Values and Units Test: attr 70 assert_equals: Value 'attr(data-foo ms)', where 'data-foo=3' should be valid for the property 'transition-duration'. expected "0.003s" but got "0s"
FAIL CSS Values and Units Test: attr 71 assert_equals: Value 'attr(data-foo time)', where 'data-foo=3ms' should be valid for the property 'transition-duration'. expected "0.003s" but got "0s"
FAIL CSS Values and Units Test: attr 72 assert_equals: Value 'attr(data-foo ms)', where 'data-foo=3' should be valid for the property 'transition-duration'. expected "0.003s" but got "0s"
PASS CSS Values and Units Test: attr 73
PASS CSS Values and Units Test: attr 74
PASS CSS Values and Units Test: attr 75
PASS CSS Values and Units Test: attr 76
FAIL CSS Values and Units Test: attr 77 assert_equals: Value 'attr(data-foo flex)', where 'data-foo=3fr' should be valid for the property 'grid-template-columns'. expected "3fr" but got "none"
FAIL CSS Values and Units Test: attr 78 assert_equals: Value 'attr(data-foo fr)', where 'data-foo=3' should be valid for the property 'grid-template-columns'. expected "3fr" but got "none"
PASS CSS Values and Units Test: attr 79
PASS CSS Values and Units Test: attr 80
PASS CSS Values and Units Test: attr 81
PASS CSS Values and Units Test: attr 82
PASS CSS Values and Units Test: attr 83
PASS CSS Values and Units Test: attr 84
PASS CSS Values and Units Test: attr 85
PASS CSS Values and Units Test: attr 86
PASS CSS Values and Units Test: attr 87
PASS CSS Values and Units Test: attr 88
PASS CSS Values and Units Test: attr 89
PASS CSS Values and Units Test: attr 90
PASS CSS Values and Units Test: attr 91
PASS CSS Values and Units Test: attr 92
PASS CSS Values and Units Test: attr 93
PASS CSS Values and Units Test: attr 94
PASS CSS Values and Units Test: attr 95
PASS CSS Values and Units Test: attr 96
PASS CSS Values and Units Test: attr 97
PASS CSS Values and Units Test: attr 98
PASS CSS Values and Units Test: attr 99
PASS CSS Values and Units Test: attr 100
PASS CSS Values and Units Test: attr 101
PASS CSS Values and Units Test: attr 102
PASS CSS Values and Units Test: attr 103
PASS CSS Values and Units Test: attr 104
PASS CSS Values and Units Test: attr 105
PASS CSS Values and Units Test: attr 106
PASS CSS Values and Units Test: attr 107
PASS CSS Values and Units Test: attr 108
